+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ### Added
 - Added filter functionality to workshops
+- Added the order functionality by name
 
+### Fixed
+- Fixed name variables for errors in workshops.model.js
 ## [0.1.0] - 10/11/2025
 
 ### Added